requirements.txt
streamlit
streamlit_extras
openai
tenacity
PyPDF2
mysql-connector-python